Delta time is double
Multiplying movement by delta time
Modified speeds
Currently still working with delta time fixed to 60 FPS

+++ src/main/java/rs/etf/dz1/sprites/Player.java
@@ -102,7 +102,7 @@ public class Player extends Sprite implements EventHandler<KeyEvent> {
     }
 
     @Override
-    public void update(long deltaTime) {
+    public void update(double deltaTime) {
         state.move();
     }
